Update the folder structure. Add index.html, images, style.css, logic.js
