+++ index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<title>Document</title>
+	<link rel="stylesheet" href="style.css">
+	<script src="style_logic.js"></script>
+</head>
+<body>
+	<!-- 	<h1 style="color:#000;font-family:sans-serif;text-align:center;">S I M P L Y &nbsp; M E</h1> -->
+	
+	<h1 style="color:#000;font-family:sans-serif;text-align:left;">P I X P R I E N C E &nbsp;</h1>
+		
+	<button style="float: right">Login</button>
+		
+</body>
+</html>
