+++ public/assets/css/style.css
@@ -1,0 +1,63 @@
+body {
+
+}
+button {
+	background: rgba(0, 0, 0, 0);
+	border: none;
+}
+#bg1 {
+	background: url("../img/Main_Image_1.jpeg") no-repeat 50% 50% fixed;
+	background-size: cover;
+	-moz-background-size: cover;
+	-o-background-size: cover;
+	-webkit-background-size: cover;
+}
+#bg2 {
+	background: url("../img/Main_Image_2.jpeg") no-repeat 50% 50% fixed;
+	background-size: cover;
+	-moz-background-size: cover;
+	-o-background-size: cover;
+	-webkit-background-size: cover;
+}
+#bg3 {
+	background: url("../img/Main_Image_3.jpeg") no-repeat 50% 50% fixed;
+	background-size: cover;
+	-moz-background-size: cover;
+	-o-background-size: cover;
+	-webkit-background-size: cover;
+}
+#bg4 {
+	background: url("../img/Main_Image_4.jpeg") no-repeat 50% 50% fixed;
+	background-size: cover;
+	-moz-background-size: cover;
+	-o-background-size: cover;
+	-webkit-background-size: cover;
+}
+#bg5 {
+	background: url("../img/Main_Image_5.jpeg") no-repeat 50% 50% fixed;
+	background-size: cover;
+	-moz-background-size: cover;
+	-o-background-size: cover;
+	-webkit-background-size: cover;
+}
+#bg6 {
+	background: url("../img/Main_Image_6.jpeg") no-repeat 50% 50% fixed;
+	background-size: cover;
+	-moz-background-size: cover;
+	-o-background-size: cover;
+	-webkit-background-size: cover;
+}
+#bg7 {
+	background: url("../img/Main_Image_7.jpeg") no-repeat 50% 50% fixed;
+	background-size: cover;
+	-moz-background-size: cover;
+	-o-background-size: cover;
+	-webkit-background-size: cover;
+}
+#bg8 {
+	background: url("../img/Main_Image_8.jpeg") no-repeat 50% 50% fixed;
+	background-size: cover;
+	-moz-background-size: cover;
+	-o-background-size: cover;
+	-webkit-background-size: cover;
+}
