--- views/index.handlebars
@@ -1,1 +1,0 @@
-<h1>{{item}}</h1>
